feature: support to load template service
When someone starts test@1.service, we first try to search test@1.service,
if we found something, we use it; if not, we will check if the service name
contains @, if yes, this means it's a template service, we then try to
search test@.service.

This commit also split the main logic of build_id_fragment to build_id_fragment_by_name.