fix(libbasic): create relative symlink in right way
When creating relative symlink, 'symlinkat' should take the directory
fd as input, but previously the 'symlink' function tried to open the
symlink that hasn't been created. The unit test case for 'symlink' just
consider the usage on creating non-relative symlink, thus it didn't
find the bug when creating relative symlink.