fix(libbasic): try to follow the fd path to make the utimensat work
In devmaster threads, utimensat will fail with errno of ENOENT, which means
the file path does not exist, if the fd path is used. The fd path is a symlink to
the opened real file. It is weired because the fd path really exists but utimesat
can not find it. To avoid the failure, we have no choice but to follow the fd path
symlink to the real file, and update the timestamp directly on it.