fix(devmaster): deal with fd leak
RawFd will not be automatically closed when it is dropped. Use std::fs::File
or nix::dir::Dir to take the RawFd when opening the file or directory, which
can automatically close the RawFd when the variable is dropped.

Also rename the fd type from i32 to RawFd to better declare the file descriptor
type.