refactor: don't copy unit file to .toml file when loading a unit
We used to do this because we thought confique could only infer
the file format from the file extension, so we copied the file
to .toml to tell confique this is a toml file. But actually,
confique allows us to specify the file format if we use Partial.

Let's use confique::Partial to avoid copying files.