feature(devmaster): support loading and applying network interface configurations
The netif configurations are loaded by devmaster and shared among
workers. In the builtin command net_setup_link, the netif configurations
are readed and applied for generating a new netif name.

This commit only implements the framework and the basic ability that
choose the original property to set the 'ID_NET_NAME' accoring to the
naming policy in the configuration.