fix: add missing features for some crates and declare feature dependencies explicitly in crate basic
When building some a crate individually, the process will abort if some required features
are not explicitly set and the default feature configuration is disabled. This patch did
two major things:

1. Declare dependencies among inner features in 'basic' crate

The indirectly depended features won't be automatically set if the dependencies are
not explicitly declared in the Cargo.toml. Thus this patch completed the dependency
declaration among features in 'basic' crate.

2. Complete missing inter feature dependencies in a series of crates

The features of external crates are completed obeying the minimal principle. That is
to say, each crate will try to disable default features of external crates and only
set the features it really needs.