feature(devmaster): notify pid 1 when devmaster is ready
If devmaster daemon is launched by sysMaster or systemd in 'notify' type
, devmaster should tell the service manager through a socket named 'NOTIFY_SOCKET'
when devmaster is ready. Then the service manager can update the state
of devmaster to activated. This mechanism lets the service manager
know at which time the devmaster is ready and try to start other services
that depend on devmaster.