fix: add open_when_needed feature for FileLogger
When we reinit the logger after closing all fds, the newly opened log fd takes up the position
that keep_fds will use, so keep_fds can't be shifted to the right position in shift_fds(), and
this leads to a dead loop.

This commit adds open_when_needed feature to FileLogger, we will only open log file when needed
and close it after logging, so the log fd won't steal the fd number position from keep_fds.