fix(devmaster): discard devctl test command that is deprecated
The 'devctl test' command is used to simulate an uevent to debug the
the framework of devmaster. After 'devctl trigger' is supported, there
is no need to use 'devctl test' anymore. Moreover, as devmaster makes
stricter uevent filter condition, 'devctl test' can not take effect
indeed currently.

If the 'devctl test' is required to be consistent with 'udevadm test'
command in future, then we will consider to implement it again.