fix: add stop_when_bound_queue fix BindsTo issue
If foo.service configured 'BindsTo=bar.service', when we start foo.service, sysmaster will automatically
start bar.service. foo.service is started first, then bar.service. If foo.service has been set to
active state, but bar.service fail to start. We should add a stop job to stop foo.service to make its
state be consistent with bar.service (This is what BindsTo mean).

This commit save foo.service in stop_when_bound_queue, and dispatch this queue in main_loop(). When dispatching
the queue, we try to find if there is a bar.service, and its state is inactive. If so, we stop foo.service