fix: don't check the event.name
event.name returns the exact name of created file in the directory
we are watching in openEuler 22.03. But in openEuler 23.09, this is not
true, event.name is not the same as the file name. The file name is
prefixed with '.#' and suffixed with random 16 digits in hexadecimal.

So we don't check the event.name here, we will check if the file exists
anyway in the beginning of the loop.